requirements.txt
pip ==24.0
setuptools ==65.5.1
darkdetect~=0.8.0
packaging~=24.0
customtkinter~=5.2.2
mysql-connector-python~=8.4.0